Moved Title/Deal Buttons, Updated and fixed background img, Centered program
I moved the title and deal button to the right side of the program. I added a background image and fixed it to the page so that when resizing the page or zooming, it wont get distorted.  I also centered the program so that it was easier to look at.