OnlyAcesCanMove
I have fixed the game to where only aces can be moved to an empty column.  I have also added java Pop-up errors when the user tries to do the following:
--Tries to move any card (unless an ace) to an empty column
--Tries to move any card to a column that isn't empty

I have also added a "Welcome" pop-up when you first start the game.  This is intended to open java on the pc (which takes a second or two), then allowing all future error messages to pop up much quicker.