Fixed ace cards
Changed Ace cards in the spanish deck to have value 13 instead of 1. Made it so jokers cannot remove aces, added tests to remove aces with joker present, added test to make sure dealFour function was only dealing 2 when there were only 2 left in the deck. Line coverage in testing is up to 99%.